+++ R/args-pkg.R
@@ -1,0 +1,11 @@
+#' @title
+#' Broca Arguments
+#'
+#' @param data          A dataframe or tibble.
+#' @param log           (Deprecated) Log this function call to a log.md file?
+#' @param log_details   (Deprecated) If log is TRUE, additional notes to include with the log entry.
+#' @param header        True if data has a header.
+#' @param quote         Single character used to quote strings.
+#'
+#' @name args
+NULL