+++ NEWS.md
@@ -1,3 +1,10 @@
+# broca 0.2.0  
+
+* Deprecate log feature  
+* Rename makeTribble -> make_trbl  
+* Add `read_cb_*` shortcut functions for reading from clipboard  
+* Add examples for clipboard function family  
+
 # broca 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* First commit